fix: improve FTA score sim-cloud-front-distribution-configurator
