+++ src/service/cloudfront/distribution/configurator/sim-cloud-front-behavior-configurator.ts
@@ -5,8 +5,8 @@ import type {
 } from "../../command/create-distribution/create-distribution.cmd.js";
 import type { SimCloudFrontBehavior } from "../../behaviour/sim-cloud-front-behavior.js";
 import { assertDefined } from "../../../../util/type-guard/defined.js";
-import type { SimArn } from "../../../aws/arn.js";
 import type { SimCloudFrontDistribution } from "../sim-cloudfront-distribution.js";
+import { configureCffAssociations } from "./sim-cff-associations-configure.js";
 
 /**
  * Applies Cache Behavior configuration to a sim CloudFront Distribution.
@@ -19,9 +19,8 @@ export class SimCloudFrontBehaviorConfigurator {
     distribution: SimCloudFrontDistribution,
     cacheBehavior: SimCloudFrontDefaultCacheBehaviorConfig | undefined,
   ): void {
-    distribution.addBehavior(
-      this.behaviorFromDefaultCacheBehavior(cacheBehavior),
-    );
+    assertDefined(cacheBehavior, "CloudFront DefaultCacheBehavior");
+    distribution.addBehavior(buildBaseBehaviorProps(cacheBehavior));
   }
 
   /**
@@ -31,117 +30,52 @@ export class SimCloudFrontBehaviorConfigurator {
     distribution: SimCloudFrontDistribution,
     cacheBehavior: SimCloudFrontCacheBehaviorConfig,
   ): void {
-    distribution.addBehavior(this.behaviorFromCacheBehavior(cacheBehavior));
-  }
-
-  private behaviorFromDefaultCacheBehavior(
-    cacheBehavior: SimCloudFrontDefaultCacheBehaviorConfig | undefined,
-  ): SimCloudFrontBehavior {
-    assertDefined(cacheBehavior, "CloudFront DefaultCacheBehavior");
-
-    return {
-      targetOriginName: this.requiredTargetOriginId(
-        cacheBehavior.TargetOriginId,
-      ),
-      allowedMethods: this.methods(cacheBehavior.AllowedMethods, [
-        "GET",
-        "HEAD",
-      ]),
-      cachedMethods: this.methods(cacheBehavior.AllowedMethods?.CachedMethods, [
-        "GET",
-        "HEAD",
-      ]),
-      ...(cacheBehavior.ViewerProtocolPolicy === undefined
-        ? {}
-        : { viewerProtocolPolicy: cacheBehavior.ViewerProtocolPolicy }),
-      functionAssociations:
-        this.functionAssociationsFromCacheBehavior(cacheBehavior),
-    };
-  }
-
-  private behaviorFromCacheBehavior(
-    cacheBehavior: SimCloudFrontCacheBehaviorConfig,
-  ): SimCloudFrontBehavior {
     assertDefined(
       cacheBehavior.PathPattern,
       "CloudFront CacheBehavior PathPattern",
     );
-
-    return {
+    distribution.addBehavior({
       pathPattern: cacheBehavior.PathPattern,
-      targetOriginName: this.requiredTargetOriginId(
-        cacheBehavior.TargetOriginId,
-      ),
-      allowedMethods: this.methods(cacheBehavior.AllowedMethods, [
-        "GET",
-        "HEAD",
-      ]),
-      cachedMethods: this.methods(cacheBehavior.AllowedMethods?.CachedMethods, [
-        "GET",
-        "HEAD",
-      ]),
-      ...(cacheBehavior.ViewerProtocolPolicy === undefined
-        ? {}
-        : { viewerProtocolPolicy: cacheBehavior.ViewerProtocolPolicy }),
-      functionAssociations:
-        this.functionAssociationsFromCacheBehavior(cacheBehavior),
-    };
+      ...buildBaseBehaviorProps(cacheBehavior),
+    });
   }
+}
 
-  private functionAssociationsFromCacheBehavior(
-    cacheBehavior:
-      | SimCloudFrontDefaultCacheBehaviorConfig
-      | SimCloudFrontCacheBehaviorConfig,
-  ): SimCloudFrontBehavior["functionAssociations"] | undefined {
-    if (cacheBehavior.FunctionAssociations?.Items === undefined) {
-      return undefined;
-    }
-    const associations: SimCloudFrontBehavior["functionAssociations"] = {};
+/**
+ * Build the shared Behavior properties common to both default and named Cache
+ * Behaviors. Callers add their own fields (e.g. pathPattern) on top.
+ */
+function buildBaseBehaviorProps(
+  cacheBehavior:
+    | SimCloudFrontDefaultCacheBehaviorConfig
+    | SimCloudFrontCacheBehaviorConfig,
+): SimCloudFrontBehavior {
+  assertDefined(
+    cacheBehavior.TargetOriginId,
+    "CloudFront CacheBehavior TargetOriginId",
+  );
 
-    for (const funcAssoc of cacheBehavior.FunctionAssociations.Items) {
-      assertDefined(
-        funcAssoc.EventType,
-        "CloudFront Function association EventType",
-      );
-      assertDefined(
-        funcAssoc.FunctionARN,
-        "CloudFront Function association FunctionARN",
-      );
-      switch (funcAssoc.EventType) {
-        case "viewer-request": {
-          associations.viewerRequest = funcAssoc.FunctionARN as SimArn;
-          break;
-        }
-        case "viewer-response": {
-          associations.viewerResponse = funcAssoc.FunctionARN as SimArn;
-          break;
-        }
-        case "origin-request": {
-          throw new Error(
-            "CloudFront Function association EventType origin-request not implemented",
-          );
-        }
-        case "origin-response": {
-          throw new Error(
-            "CloudFront Function association EventType origin-response not implemented",
-          );
-        }
-      }
-    }
+  return {
+    targetOriginName: cacheBehavior.TargetOriginId,
+    allowedMethods: methodsSet(cacheBehavior.AllowedMethods, ["GET", "HEAD"]),
+    cachedMethods: methodsSet(cacheBehavior.AllowedMethods?.CachedMethods, [
+      "GET",
+      "HEAD",
+    ]),
+    ...(cacheBehavior.ViewerProtocolPolicy === undefined
+      ? {}
+      : { viewerProtocolPolicy: cacheBehavior.ViewerProtocolPolicy }),
+    functionAssociations: configureCffAssociations(cacheBehavior),
+  };
+}
 
-    return Object.keys(associations).length > 0 ? associations : undefined;
-  }
-
-  private requiredTargetOriginId(targetOriginId: string | undefined): string {
-    assertDefined(targetOriginId, "CloudFront CacheBehavior TargetOriginId");
-
-    return targetOriginId;
-  }
-
-  private methods(
-    methods: SimCloudFrontMethodList | undefined,
-    fallback: string[],
-  ): Set<string> {
-    return new Set(methods?.Items ?? fallback);
-  }
+/**
+ * Build a Set of HTTP methods from a method list config, falling back to the
+ * provided defaults when no items are configured.
+ */
+function methodsSet(
+  methods: SimCloudFrontMethodList | undefined,
+  fallback: string[],
+): Set<string> {
+  return new Set(methods?.Items ?? fallback);
 }
